MAPR-22888 Hive-2.0.0 ERROR No such file or directory after querying data from any table via HiveCLI

+++ ql/src/java/org/apache/hadoop/hive/ql/Context.java
@@ -376,8 +376,10 @@ public class Context {
       try {
         Path p = entry.getValue();
         FileSystem fs = p.getFileSystem(conf);
-        fs.delete(p, true);
-        fs.cancelDeleteOnExit(p);
+        if(fs.exists(p)) {
+          fs.delete(p, true);
+          fs.cancelDeleteOnExit(p);
+        }
       } catch (Exception e) {
         LOG.warn("Error Removing Scratch: "
             + StringUtils.stringifyException(e));
